refactors update card acceptance tests
