+++ src/test/resources/features/updateCards.feature
@@ -1,7 +1,6 @@
-Feature: card is edited
+Feature: updating card
 
-  Scenario: client makes call to PATCH /cards/1
+  Scenario: client updates a card
     Given the card exists
-    When the client updates to cards with this id and changes the text to "bye"
-    Then the client receives 200 status code
-    And the client receives the card with the text:"bye"
+    When the client updates the existing card with text "bye"
+    Then the client receives the card with the text:"bye"
